Extracted parser from component into separate module.
